Revert "Basic functionality for declaring smt-lemmas"
This reverts commit 73197db0c0165dff22746013c9ac4ac4ca2166dc.
